Deleting apps now deletes associated images and data

+++ cloudComputeCostWaste/models/app.js
@@ -1,5 +1,4 @@
 var mongoose = require('mongoose');
-var User = mongoose.model('User') 
 
 // User Schema
 var AppSchema = mongoose.Schema({
@@ -11,6 +10,9 @@ var App = module.exports = mongoose.model('App', AppSchema);
 
 module.exports.addData = function (name, user, done)
 {
+	var User = mongoose.model('User');
+	var Image = mongoose.model('Image');
+	var Data = mongoose.model('Data');
 	User.findOne({ _id: user }, function(err, userRef) {
 	    if(err) {
 	        console.log(err);  // handle errors!
@@ -36,7 +38,25 @@ module.exports.addData = function (name, user, done)
 
 module.exports.delete = function (id, callback)
 {
+	var User = mongoose.model('User');
+	var Image = mongoose.model('Image');
+	var Data = mongoose.model('Data');
+	
 	var id = mongoose.Types.ObjectId(id);
-	console.log("Searching for id of app to delete");
-	App.findOneAndRemove({ _id: id }, callback);
+	Image.find({ _app: id }, function(err, images) {
+		if(err) {
+			console.log(err);
+		}
+		let itemsProcessed = 0;
+		console.log(images);
+		for(let i = 0; i < images.length; ++i) {
+			Image.delete(images[i]._id, function(){
+				++itemsProcessed;
+				if(itemsProcessed == images.length) {
+					console.log("Searching for id of app to delete");
+					App.findOneAndRemove({ _id: id }, callback);
+				}
+			});
+		}
+	});
 };
